[IMP][V2.1]split into 2 modules
previously, all features were in the module electricity contracts
some may need features of electricity contract but not the mrp
split into 2 modules (add electricity_mrp, dpending of electricity_contract)
